Added next repos and rest services: -Waybill rest -RouteList rest (PUT doesn't work with Truck entities{Truck doesn't exist}) -WriteOffAct rest

+++ src/main/java/com/truckcompany/repository/WriteOffActRepository.java
@@ -9,8 +9,8 @@ import java.util.Optional;
 /**
  * Created by Viktor Dobroselsky.
  */
-public interface WriteOffActRepository extends JpaRepository<WriteOffAct, Integer> {
+public interface WriteOffActRepository extends JpaRepository<WriteOffAct, Long> {
     List<WriteOffAct> findAll ();
 
-    Optional<WriteOffAct> findOneById(Integer id);
+    Optional<WriteOffAct> findOneById(Long id);
 }
